+++ source/Energinet.DataHub.MeteringPoints.IntegrationTests/UpdateMasterData/TestUtils.cs
@@ -15,7 +15,7 @@
 using Energinet.DataHub.MeteringPoints.Application.MarketDocuments;
 using Energinet.DataHub.MeteringPoints.Domain;
 
-namespace Energinet.DataHub.MeteringPoints.IntegrationTests.UpdateMasterData.ConsumptionMeteringPoints
+namespace Energinet.DataHub.MeteringPoints.IntegrationTests.UpdateMasterData
 {
     public static class TestUtils
     {
